Add package liblsl explicitly to Mahalia image
liblsl is currently installed in the Mahalia image because
openMHA has a "Depends:" dependency on it.  T1731 plans to
change this dependency into a "Recommends:" dependency.
In order to make sure that liblsl still gets installed
in the Mahalia image, we need to list liblsl in the list
of packages installed from our apt server explicitly.

Closes T1727.